Move predict models to internal module
