+++ nucliadb_models/src/nucliadb_models/internal/predict.py
@@ -1,0 +1,65 @@
+# Copyright (C) 2021 Bosutech XXI S.L.
+#
+# nucliadb is offered under the AGPL v3.0 and as commercial software.
+# For commercial licensing, contact us at info@nuclia.com.
+#
+# AGPL:
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+#
+
+"""
+Models for Predict API v1.
+
+ATENTION! Keep these models in sync with models on Predict API
+"""
+
+from typing import List, Optional
+
+from pydantic import BaseModel, Field
+
+
+class SentenceSearch(BaseModel):
+    data: List[float] = Field(deprecated=True, default_factory=list)
+    vectors: dict[str, List[float]] = Field(description="Sentence vectors for each semantic model")
+    time: float = Field(deprecated=True)
+    timings: dict[str, float] = Field(
+        description="Time taken to compute the sentence vector for each semantic model"
+    )
+
+
+class Ner(BaseModel):
+    text: str
+    ner: str
+    start: int
+    end: int
+
+
+class TokenSearch(BaseModel):
+    tokens: List[Ner] = []
+    time: float
+    input_tokens: int = 0
+
+
+class QueryInfo(BaseModel):
+    language: str
+    stop_words: List[str]
+    semantic_threshold: float = Field(deprecated=True)
+    semantic_thresholds: dict[str, float] = Field(
+        description="Semantic threshold for each semantic model"
+    )
+    visual_llm: bool
+    max_context: int
+    entities: Optional[TokenSearch]
+    sentence: Optional[SentenceSearch]
+    query: str
